requirements.txt
pandas
fpl-data-loader
streamlit>=1.35.0
requests
plotly